+++ src/Microsoft.ApplicationInsights.AspNetCore/DiagnosticListeners/Implementation/IApplicationInsightDiagnosticListener.cs
@@ -1,11 +1,12 @@
 ﻿namespace Microsoft.ApplicationInsights.AspNetCore.DiagnosticListeners
 {
     using System;
+    using System.Collections.Generic;
 
     /// <summary>
     /// Base diagnostic listener type for Application Insight
     /// </summary>
-    internal interface IApplicationInsightDiagnosticListener : IDisposable
+    internal interface IApplicationInsightDiagnosticListener : IDisposable, IObserver<KeyValuePair<string, object>>
     {
         /// <summary>
         /// Gets a value indicating which listener this instance should be subscribed to